refactor: extract the pty spawn primitives and draft injection (#548 step 3b)
Two groups come out of index.ts ahead of the spawn functions built on them,
because neither touches index.ts's session state:

- session/pty-spawn.ts — spawnPty, ptySpawn, spawnSandboxEntry, sandboxWouldRun.
  The ways a terminal gets started (bare, tmux-backed, Docker sandbox); they
  depend only on infra.
- session/draft-injection.ts — sanitizeDraftText and the two readiness strategies
  that type a prompt into an agent's TUI (claude matches a marker, codex waits
  for its output to settle).

Every moved body is byte-identical to its original. sanitizeDraftText is pure and
now tested: it is the guard between untrusted text and a live terminal, so its
job is to leave no control byte that could close the bracketed paste early.
